+++ lib/mirage_configure.ml
@@ -63,9 +63,9 @@ let find_git () =
   Bos.OS.Dir.current () >>= fun cwd ->
   find cwd None >>= fun subdir ->
   let git_branch = Bos.Cmd.(v "git" % "rev-parse" % "--abbrev-ref" % "HEAD") in
-  Bos.OS.Cmd.(run_out git_branch |> out_string) >>= fun (branch, _) ->
+  Bos.OS.Cmd.(run_out git_branch |> out_string |> success) >>= fun branch ->
   let git_remote = Bos.Cmd.(v "git" % "remote" % "get-url" % "origin") in
-  Bos.OS.Cmd.(run_out git_remote |> out_string) >>| fun (git_url, _) ->
+  Bos.OS.Cmd.(run_out git_remote |> out_string |> success) >>| fun git_url ->
   subdir, branch, git_url
 
 let configure_opam ~name info =